Update VMCImportExport.py
Adding ability to skip import of groups based on manual VM membership (as these are based on VM external ID, they will not be restored as part of the import/export). Without this, the tool would restore the configuration of the group but the group would be empty or incorrectly implemented (as the group would use the "source" SDDC's external ID of the VM, which is no longer valid at the destination).